# Improve performance of fetching a cached channel by avoiding a full iteration