This PR adds an overload to InternalGetCachedChannel which takes a guild ID, which  skips iterating ALL guilds to find a channel. There are some instances where the guild ID is not available, and thus iteration must prevail, but in 90% of the API calls this is present, and thus can benefit.

The sample data (and thus, PR title) is derived from testing with about 4000 dictionary keys

```fix
| Method               | Mean          | Error       | StdDev      |
|--------------------- |--------------:|------------:|------------:|
| WorstCaseIterate     | 19,867.684 ns | 326.9402 ns | 273.0099 ns |
| BestCaseIterate      |      7.253 ns |   0.1474 ns |   0.2251 ns |
| WorstCaseTryGetValue |      3.941 ns |   0.0711 ns |   0.0631 ns |
| BestCaseTryGetValue  |      6.927 ns |   0.0987 ns |   0.0825 ns |
```

It is also worth noting that "Worst" and "Best" for TGV are actually backward - Worst case is 1 call (guild not present) and "best" case is when the guild and channel are present. 

---

Reviewer's note: if you ever dispute that this is faster than the code it replaces, you should get your performance senses checked, but since there's a fair bit more to it than the benchmarks here, see the discussion on discord: https://canary.discord.com/channels/379378609942560770/379386901725052928/1265540956816412734