@VelvetToroyashi 2024-07-24
Running D#+ under a profiler prompted me to fix this issue, because 1.5 seconds (across the profiler's runtime, I think) is...*egregious*. Also, take my benchmark with a grain of salt, because that's a *normal* Dictionary, but D#+ uses a *ConcurrentDictionary*, which will probably only be even worse for iterating values over.. 
![image](https://github.com/user-attachments/assets/605f5c6e-9bbe-4076-8f46-5bf0c4ca3943)