@VelvetToroyashi 2024-07-24
ConcurrentDictionary makes me sad ![:catresort:](https://cdn.discordapp.com/emojis/1044308251455533136.webp?size=24&quality=lossless)

Difference is about 57000% (5666 repeating times better, give or take)

```fix
| Method               | Mean          | Error       | StdDev      |
|--------------------- |--------------:|------------:|------------:|
| WorstCaseIterate     | 33,956.363 ns | 323.7009 ns | 252.7243 ns |
| BestCaseIterate      | 14,778.750 ns | 228.1886 ns | 202.2833 ns |
| WorstCaseTryGetValue |      6.191 ns |   0.1063 ns |   0.0994 ns |
| BestCaseTryGetValue  |     17.967 ns |   0.3873 ns |   0.3978 ns |
```